@@ -1,49 +1,23 @@
-[package]
-name = "externref"
-version = "0.1.0"
-authors = ["Alex Ostrovski <ostrovski.alex@gmail.com>"]
+[workspace]
+members = ["crates/lib", "crates/macro", "e2e-tests"]
+
+[workspace.package]
+version = "0.2.0"
 edition = "2021"
-rust-version = "1.64"
-readme = "README.md"
+rust-version = "1.66"
+authors = ["Alex Ostrovski <ostrovski.alex@gmail.com>"]
 license = "MIT OR Apache-2.0"
-keywords = ["externref", "anyref", "wasm"]
-categories = ["wasm", "development-tools::ffi"]
-description = "Low-cost reference type shims for WASM modules"
 repository = "https://github.com/slowli/externref"
 
-[package.metadata.docs.rs]
-all-features = true
-# Set `docsrs` to enable unstable `doc(cfg(...))` attributes.
-rustdoc-args = ["--cfg", "docsrs"]
-
-[badges]
-maintenance = { status = "experimental" }
-
-[dependencies]
-externref-macro = { version = "0.1.0", path = "crates/macro", optional = true }
-# Processor dependencies
-anyhow = { version = "1.0.58", optional = true }
-walrus = { version = "0.19.0", optional = true }
-# Enables tracing during module processing
-tracing = { version = "0.1.37", optional = true }
-
-[dev-dependencies]
+[workspace.dependencies]
+# General-purpose dependencies
+anyhow = "1.0.66"
+once_cell = "1.13.0"
+tracing = "0.1.37"
+# Test dependencies
 assert_matches = "1.5.0"
 doc-comment = "0.3.3"
 version-sync = "0.9.4"
-wat = "1.0.46"
-
-[features]
-default = ["macro"]
-# Re-exports the `externref` macro
-macro = ["externref-macro"]
-# Enables WASM module processing logic (the `processor` module)
-processor = ["anyhow", "walrus"]
-
-[[test]]
-name = "processor"
-path = "tests/processor.rs"
-required-features = ["processor"]
 
 # Profile for WASM modules generated in E2E tests
 [profile.wasm]
@@ -52,6 +26,3 @@ panic = "abort"
 codegen-units = 1
 opt-level = "z" # Optimize for size, rather than speed
 lto = true
-
-[workspace]
-members = [".", "crates/macro", "e2e-tests"]
